requirements.txt
Django>=3.0.2,<3.1.0
djangorestframework>=3.10.0,<3.11.0
python-dateutil>=2.6.1,<2.7.0
dask>=2.10.0,<2.11.0
distributed>=2.10.0,<2.11.0
bokeh>=1.4.0,<1.5.0
streamz>=0.5.2,<0.6.0
phonenumbers>=8.11.2,<8.12.0
googletrans>=2.4.0,<2.5.0
lxml>=4.4.2,<4.5.0
PyDictionary>=1.5.2,<1.6.0
requests>=2.22.0,<2.23.0